% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eval.estimation}
\alias{eval.estimation}
\title{Parameter Estimation Evaluation.}
\usage{
eval.estimation(rdist, edist, n = 20, rep.num = 1000, params,
  method = "numerical.MLE")
}
\arguments{
\item{rdist}{random variable generating function.}

\item{edist}{parameter estimation function.}

\item{n}{sample size}

\item{rep.num}{number of replacates}

\item{params}{true parameters of the distribution.}

\item{method}{esimation method}
}
\value{
a list containing the mean, sd of the estimated parameters. na.cont
is the number of "na"s appeared in the parameter estimation.
}
\description{
A function to evaluate the parameter estimation function.
}
\details{
Parameter Estimation Evaluation

Details
}
\examples{
\donttest{
eval.estimation(rdist=rNormal,edist=eNormal,n = 10, rep.num = 1e3,
params = list(mean = 1,sd = 5))
eval.estimation(rdist=rNormal,edist=eNormal,n = 1000, rep.num = 1e3,
params = list(mean = 1,sd = 5))
}
}
\author{
Haizhen Wu and A. Jonathan R. Godfrey
}

