% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmle.R
\name{wmle}
\alias{wmle}
\title{Weighted Maximum Likelihood Estimation.}
\usage{
wmle(X, w, distname, initial, lower, upper, loglik.fn, score.fn, obs.info.fn)
}
\arguments{
\item{X}{Sample observations.}

\item{w}{Frequency (or weights) of observation.}

\item{distname}{Name of distribution to be estimated.}

\item{initial}{Initial value of the parameters.}

\item{lower}{The lower bound of the parameters.}

\item{upper}{The upper bound of the parameters.}

\item{loglik.fn}{Function to compute (weighted) log likelihood.}

\item{score.fn}{Function to compute (weighted) score.}

\item{obs.info.fn}{Function to compute observed information matrix.}
}
\value{
weighted mle estimates.
}
\description{
A general weighted maximum likelihood estimation function.
}
\details{
Weighted Maximum Likelihood Estimation
}
\author{
Haizhen Wu and A. Jonathan R. Godfrey
}
