% Generated by roxygen2 (4.0.0): do not edit by hand
\name{compareDist}
\alias{compareDist}
\title{Compare sample and fitted distributions}
\usage{
compareDist(X, Dist1, Dist2 = NULL, Dist3 = NULL)
}
\arguments{
\item{X}{- the (unweighted) sample}

\item{Dist1,Dist2,Dist3}{- the eDist objects or names of the distribution to be fitted.}
}
\description{
A method to Compare sample and fitted distributions
}
\examples{
\donttest{
n <- 100
par <- list(shape1=1, shape2=2)
X <- rBeta(n, params=par)
compareDist(X, "Beta", "Normal")
compareDist(X, "Beta", eNormal(X))
}
}
\author{
Haizhen Wu and A. Jonathan R. Godfrey
}

