% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.estimation.R
\name{eval.estimation}
\alias{eval.estimation}
\title{Parameter Estimation Evaluation.}
\usage{
eval.estimation(
  rdist,
  edist,
  n = 20,
  rep.num = 1000,
  params,
  method = "numerical.MLE"
)
}
\arguments{
\item{rdist}{Random variable generating function.}

\item{edist}{Parameter estimation function.}

\item{n}{Sample size.}

\item{rep.num}{Number of replicates.}

\item{params}{True parameters of the distribution.}

\item{method}{Estimation method.}
}
\value{
A list containing the mean and sd of the estimated parameters.\cr
\cr
 na.cont returns the number of "na"s that appeared in the parameter estimation.
}
\description{
A function to evaluate the parameter estimation function.
}
\examples{
eval.estimation(rdist = rBeta, edist = eBeta, n = 100, rep.num = 50,
params = list(shape1 = 1, shape2 = 5))
}
\author{
Haizhen Wu and A. Jonathan R. Godfrey.
}
