\name{ExtDist-package}
\alias{ExtDist-package}
\alias{ExtDist}
\docType{package}
\title{
Extended Probability Distribution Functions
}
\description{
The package provides a consistent, unified and extensible
    framework for parameter estimation of probability distributions; it extends
    parameter estimation procedures to allow for weighted samples; moreover, it
    extends the gallery of available distributions.
}
\details{
\tabular{ll}{
Package: \tab ExtDist\cr
Version: \tab 0.3.3\cr
License: \tab GPL (>= 2)\cr
Imports: \tab 
numDeriv,
optimx,
matrixcalc\cr
Suggests: \tab 
knitr\cr
Roxygen: \tab list(wrap = FALSE)\cr
VignetteBuilder: \tab knitr\cr
Built: \tab R 3.1.0; ; 2014-07-11 04:55:47 UTC; unix\cr
}

Index:
\preformatted{
Beta                    The standard Beta Distribution.
Beta.ab                 The four-Parameter beta Distribution.
ExtDist-package         Extended Probability Distribution Functions
Laplace                 The Laplace Distribution.
Normal                  The Normal Distribution.
compareDist             Compare sample and fitted distributions
eDist                   S3 methods from manuplating eDist objects.
eval.estimation         Parameter Estimation Evaluation.
wmle                    Weighted Maximum Likelihood Estimation.
}

Further information is available in the following vignettes:
\tabular{ll}{
\code{Distributions-Beta} \tab Distributions-Beta (source, pdf)\cr
\code{Distributions-Normal} \tab Distributions-Normal (source, pdf)\cr
\code{Distributions} \tab Distributions-Index (source, pdf)\cr
}
}
\author{
Haizhen Wu <h.wu2@massey.ac.nz>,
    A. Jonathan R. Godfrey <A.J.Godfrey@massey.ac.nz>,
    Kondaswamy Govindaraju <k.govindaraju@massey.ac.nz>

Maintainer: Haizhen Wu <h.wu2@massey.ac.nz>
}
